# prepareFormat for preparing format functions