This is a solution for fixing #423 

I need help on this, especially on export declarations and needed documentation. Should the function be in its own file?

Anyway, here is the benchmark code:
```ts
import Benchmark from 'benchmark';
const formatNew = require('./src/format')

const date = new Date();

let prepared = formatNew.prepareFormat("y")

const testsFunctions = {
  "old format new lib         ": {
    test: (formatStr) => formatNew.formatOld(date, formatStr)
  },
  "new format one shot prepare": {
    test: (formatStr) => formatNew.default(date, formatStr)
  },
  "new format prepare nooverhead": {
    test: (formatStr) => formatNew.prepareFormat(formatStr, {date})
  },
  "using already prepared     ": {
    init: (formatStr) => prepared = formatNew.prepareFormat(formatStr),
    test: () => prepared(date)
  }
};

const formatStrings = [
  "PPPppp",
  "P",
  "yyyyMMdd'T'HHmmss.SSS",
  "'aloa'",
  "yyyy",
]

formatStrings.forEach(formatStr => {
  const suite = new Benchmark.Suite;

  
  const output = Object.values(testsFunctions).map(test => {
    test.init?.(formatStr)
    return test.test(formatStr)
  })
  const isSame = new Set(output).size <= 1

  console.log('Output for', formatStr, 'is', output[0], isSame ? "✅" : "❌ " + output)
  Object.entries(testsFunctions).forEach(([name, test]) => suite.add(name, () => test.test(formatStr)))
  suite
    .on('cycle', (event) => console.log(String(event.target)))
    .on('complete', () => console.log(''))
    .run();
});
```

Output:
```
❯ yarn node -r ts-node/register benchmark.js
yarn node v1.22.19
Output for PPPppp is January 19th, 2023 at 3:57:46 PM GMT+1 ✅
old format new lib          x 179,903 ops/sec ±2.42% (90 runs sampled)
new format one shot prepare x 158,100 ops/sec ±3.20% (84 runs sampled)
new format prepare nooverhead x 180,275 ops/sec ±2.14% (85 runs sampled)
using already prepared      x 338,978 ops/sec ±2.91% (87 runs sampled)

Output for P is 01/19/2023 ✅
old format new lib          x 478,342 ops/sec ±4.42% (81 runs sampled)
new format one shot prepare x 425,239 ops/sec ±2.56% (80 runs sampled)
new format prepare nooverhead x 711,910 ops/sec ±4.29% (85 runs sampled)
using already prepared      x 778,732 ops/sec ±2.25% (92 runs sampled)

Output for yyyyMMdd'T'HHmmss.SSS is 20230119T155746.685 ✅
old format new lib          x 174,275 ops/sec ±6.63% (62 runs sampled)
new format one shot prepare x 221,499 ops/sec ±3.95% (84 runs sampled)
new format prepare nooverhead x 282,716 ops/sec ±1.95% (89 runs sampled)
using already prepared      x 432,093 ops/sec ±1.43% (90 runs sampled)

Output for 'aloa' is aloa ✅
old format new lib          x 1,210,808 ops/sec ±1.46% (90 runs sampled)
new format one shot prepare x 960,371 ops/sec ±2.80% (83 runs sampled)
new format prepare nooverhead x 1,657,712 ops/sec ±3.53% (85 runs sampled)
using already prepared      x 2,717,344 ops/sec ±2.99% (91 runs sampled)

Output for yyyy is 2023 ✅
old format new lib          x 699,194 ops/sec ±4.87% (75 runs sampled)
new format one shot prepare x 755,738 ops/sec ±3.44% (91 runs sampled)
new format prepare nooverhead x 1,383,233 ops/sec ±2.69% (85 runs sampled)
using already prepared      x 1,218,078 ops/sec ±3.69% (82 runs sampled)
```