@kossnocorp 2023-12-22
Hey! I appreciate it! Can you please elaborate on what function the questions are about, the optimized format or the benchmark? There's no need for documentation, but I'll gladly have the benchmark in a repo somewhere.

If you can start by rebasing the code and opening complete PR so I can run tests, I can later review the changes and find a place for benchmarks for future references.

I hope you understand the lack of previous response. I was quite busy with preparing v3, and this (like a lot more other things) slipped through the cracks.